backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
python-multipart==0.0.20
pdfplumber==0.11.6
python-docx==1.1.2
google-genai==1.12.1
pydantic==2.11.3
python-dotenv==1.1.0
